requirements.txt
Django==3.2.5
asgiref==3.4.0
psycopg2==2.9.1
dj-database-url==0.5.0
django-heroku==0.0.0
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django-bootstrap-datepicker-plus==3.0.5